requirements.txt
einops==0.6.1
matplotlib==3.7.2
numpy==1.23.5
opencv_python==4.8.0.74
PyYAML==6.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
datasets
transformers
Pillow
